moving tests directory to top level
