+++ .travis.yml
@@ -8,7 +8,7 @@ sudo: false
 
 services:
   - postgresql
-  
+
 python:
 - '3.6'
 - '3.7'
@@ -39,7 +39,7 @@ install:
 - pip install .[all,dev]
 
 script:
-- pytest -p no:sugar python/sdssdb/tests --cov python/sdssdb --cov-report html
+- pytest -p no:sugar tests --cov python/sdssdb --cov-report html
 
 after_success:
 - coveralls
