+++ tests/pwdbs/models.py
@@ -1,31 +1,29 @@
 # !/usr/bin/env python
 # -*- coding: utf-8 -*-
-# 
+#
 # Filename: models.py
-# Project: sqladbs
+# Project: pwdbs
 # Author: Brian Cherinka
-# Created: Sunday, 1st March 2020 5:45:37 pm
+# Created: Monday, 23rd March 2020 4:35:07 pm
 # License: BSD 3-clause "New" or "Revised" License
 # Copyright (c) 2020 Brian Cherinka
-# Last Modified: Tuesday, 24th March 2020 2:33:53 pm
+# Last Modified: Tuesday, 24th March 2020 2:25:39 pm
 # Modified By: Brian Cherinka
 
 
 from __future__ import print_function, division, absolute_import
-
-from sqlalchemy import Column, Integer, String
-from sdssdb.tests.sqladbs import TmpBase
+from peewee import CharField, AutoField
+from tests.pwdbs import TmpModel
 
 #
 # This file contains models used by the temporary test database.
-# New test models can be defined by subclassing from TmpBase
+# New test models can be defined by subclassing from TmpModel
 #
 
 
-class User(TmpBase):
+class User(TmpModel):
     ''' model for user on test database '''
     __tablename__ = 'user'
-    id = Column(Integer, primary_key=True)
-    name = Column(String(80))
-    essence = Column(String(10))
-
+    id = AutoField()
+    name = CharField()
+    essence = CharField()
