+++ tests/sqladbs/conftest.py
@@ -1,13 +1,13 @@
 # !/usr/bin/env python
 # -*- coding: utf-8 -*-
-# 
+#
 # Filename: conftest.py
-# Project: pwdbs
+# Project: sqladbs
 # Author: Brian Cherinka
-# Created: Monday, 23rd March 2020 4:41:32 pm
+# Created: Sunday, 1st March 2020 5:46:55 pm
 # License: BSD 3-clause "New" or "Revised" License
 # Copyright (c) 2020 Brian Cherinka
-# Last Modified: Tuesday, 24th March 2020 2:53:13 pm
+# Last Modified: Tuesday, 24th March 2020 2:53:02 pm
 # Modified By: Brian Cherinka
 
 
@@ -16,9 +16,8 @@ import factory
 import inspect
 import pytest
 from pytest_factoryboy import register
-from . import factories
 from ..conftest import determine_scope
-
+from . import factories
 
 # register all xxxFactory classes
 #
@@ -28,16 +27,16 @@ from ..conftest import determine_scope
 # @register class decorator. For docs on pytest_factoryboy,
 # see https://pytest-factoryboy.readthedocs.io/en/latest
 for item in dir(factories):
-    if item == 'PeeweeModelFactory':
-        continue
     attr = getattr(factories, item)
     if inspect.isclass(attr) and issubclass(attr, factory.Factory):
         register(attr)
 
 
 @pytest.fixture(scope=determine_scope, autouse=True)
-def transaction(database):
-    ''' Peewee transaction fixture. set autouse=True to ensure persistence '''
-    with database.transaction() as txn:
-        yield txn
-        txn.rollback()
+def session(database):
+    ''' SQLA session fixture. set autouse=True to ensure persistence '''
+    session = database.Session()
+    session.begin()
+    yield session
+    session.rollback()
+    session.close()
