+++ tests/test_connection.py
